Automatically convert PDFs to TIFFs and upload
When --upload is given a PDF file, the pages are rendered down
to TIFFs and uploaded as individual scan files.  This permits
the user to split a document apart, such as if their scanner
produced a single giant PDF file from one large batch input.

Requires the 'pdfinfo' and 'convert' utility programs to do the
heavy lifting of decoding the PDF file and rendering its pages
down to 300 DPI TIFF images.