Cleanup document tile and description formatting in lists
Instead of using quotes around the document title, just display
it as the bullet item in bold.

Instead of showing None for the description, place it in parens
when its available, and otherwise show nothing at all.  Most of
the documents scanned don't have descriptions, because the web
UI doesn't permit this field to be set (yet).